+++ packages/file_picker_darwin/lib/src/file_picker_darwin.dart
@@ -117,6 +117,7 @@ class FilePickerDarwin extends FilePickerPlatform {
             'compressionQuality': compressionQuality,
             'assetRepresentationMode':
                 darwinOptions.assetRepresentationMode.name,
+            'acceptLabel': darwinOptions.acceptLabel,
           });
 
       if (result == null) {